requirements.txt
fastapi
uvicorn
python-multipart
pydantic
chromadb
python-dotenv
openai
ebooklib
beautifulsoup4
langchain-text-splitters
python-magic
rq
redis
chardet